major bug fix & update to descriptor
Fixed a major bug that was preventing the error object from ever having more than one array cell, and the error was being placed in the second array cell rather than the first. 

Also updated the way the descriptor works. The parameter is now called "customMessage", not "descriptor". The descriptor element in the error object will now always be the FileMaker error code descriptor. If the user passes in a customMessage, it will be in the element "message". Additionally, if there is no custom message specified, then the descriptor text will be automatically copied to the "message" element value.