fixed bug with number type
A large number would not work properly, and returned as type text. Overhauled how the function was determining number vs. text. FileMaker passes in a text "number" the same if it's actually a number, so there's no was to tell them apart unless it's a value less than one, like "0.123"if it's passed in as text, which would look like ".123" if it's passed in as a number. So I scrapped that method and instead it always returns a number as having a higher order type of text (i.e. any number could also be text)